Reworked unit tests of rename transformer to verify that both modes works the same
